pyproject.toml
[project]
name = "prospection-unijui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.1",
    "flask>=3.1.1",
    "flask-migrate>=4.1.0",
    "flask-restx>=1.3.0",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
]
[tool.pyright]
# Define your project root. Often '.' (the current directory) is sufficient.
# If your source code is in a subdirectory (e.g., 'src'), specify it here.
include = ["src"]
exclude = [
    "**/__pycache__",
    "**/.venv", 
    "**/tests", 
]

# Report missing imports - crucial for catching unresolved relative paths
reportMissingImports = true

# Report missing type stubs - useful, but can be noisy for libraries without stubs
reportMissingTypeStubs = false

venvPath = "."
venv = ".venv" 

